Crate fe_semantics

Fe semantic analysis.

This library is used to analyze the semantics of a given Fe AST. It detects any semantic errors within a given AST and produces a Context instance that can be used to query contextual information attributed to AST nodes.

Modules

errors
Semantic errors.

namespace

test_utils

Structs

Context
Contains contextual information about a Fe module and can be queried using Spanned AST nodes.

ContractAttributes
Contains contextual information relating to a contract AST node.

ExpressionAttributes
Contains contextual information relating to an expression AST node.

FunctionAttributes
Contains contextual information relating to a function definition AST node.

Enums

Location
Indicates where an expression is stored.

RuntimeOperations
Operations that need to be made available during runtime.

Functions

analysis
Performs semantic analysis of the source program and returns a Context instance.